+++ sponge_remote_mobile/CHANGELOG.md
@@ -1,8 +1,9 @@
 ## 0.16.2 (not released yet)
 
+* Added Service discovery warning if WiFi is turned off.
+* Fixed WiFi name in service discovery for Android API since 23.
 * Fixed navigation bar color issue on some Android devices.
 * Fixed event subscription animated icon.
-* Added Service discovery warning if WiFi is turned off.
   
 ## 0.16.1 (2020-07-31)